Optimized Hausdorff and ModifiedHausdorff algorithms (fixes #56). (#57)
* Optimized Hausdorff and ModifiedHausdorff algorithms (fixes #56).

Previously, the Hausdorff routines used time and memory that was
quadratic in the number of active pixels in the input images. This
made them impractical for large or dense images (see #56).

Here, we leverage the `distance_transform` from `ImageMorphology`
to compute the Hausdorff distance in linear time. However, because
this new algorithm operates on a dense representation of the images,
it may be slower for some very large and very sparse images.

This change also adds `ImageMorphology` as a dependency.